refactor: replace ContentAreaLoader with FullScreenLoader in ContentWithDrawer and ContentWithoutDrawer
