+++ frontend/src/community/common/components/organisms/ContentWithoutDrawer/ContentWithoutDrawer.tsx
@@ -1,8 +1,9 @@
 import { Stack } from "@mui/material";
 import { ReactNode } from "react";
 
-import ContentAreaLoader from "~community/common/components/molecules/ContentAreaLoader/ContentAreaLoader";
+import FullScreenLoader from "~community/common/components/molecules/FullScreenLoader/FullScreenLoader";
 import ToastMessage from "~community/common/components/molecules/ToastMessage/ToastMessage";
+import useRouteLoading from "~community/common/hooks/useRouteLoading";
 import {
   initialState,
   useToast
@@ -18,11 +19,12 @@ const ContentWithoutDrawer = ({ children }: Props) => {
   const classes = styles();
 
   const { toastMessage, setToastMessage } = useToast();
+  const loading = useRouteLoading();
 
   return (
     <>
       <Stack sx={classes.unProtectedWrapper}>
-        <ContentAreaLoader fullPage />
+        {loading && <FullScreenLoader />}
         {children}
         <ToastMessage
           key={toastMessage.key}
